Split components out into individual files
